Refactor hooks and add lib helpers
