+++ app/src/lib/math-utils.js
@@ -1,0 +1,8 @@
+export function round(num, places = 2) {
+  const rounded = Number(Math.round(num + 'e+' + places) + 'e-' + places)
+  return Number.isNaN(rounded) ? Number(num.toFixed(places)) : rounded
+}
+
+export function toHours(ms) {
+  return ms / 3600000
+}
